requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
arxiv==1.4.2
requests==2.26.0
openai==1.3.0


